Remove polars and actix-web. Implement custom parquet query. Go back to rocksdb (#170)
- Removed polars dependency and replaced it with custom code to query parquet files.
- Go back to rocksdb because of issues with libmdbx.

+++ core/src/parquet_source/ver0_0_39.rs
@@ -5,13 +5,12 @@ use super::{Columns, ParquetSource};
 use crate::deserialize::{Address, BigUnsigned, BloomFilterBytes, Bytes, Bytes32, Index};
 use crate::types::{Block, Log, Transaction};
 use arrayvec::ArrayVec;
-use polars::export::arrow::array::{self, BooleanArray, Int64Array, UInt32Array, UInt64Array};
-use polars::export::arrow::compute::cast::cast;
-use polars::export::arrow::compute::concatenate::concatenate;
-use polars::export::arrow::datatypes::{DataType, Field};
+use arrow2::array::{self, BooleanArray, Int64Array, UInt32Array, UInt64Array};
+use arrow2::compute::concatenate::concatenate;
+use arrow2::datatypes::{DataType, Field};
 use std::collections::BTreeMap;
 
-type BinaryArray = array::BinaryArray<i64>;
+type BinaryArray = array::BinaryArray<i32>;
 
 pub struct Ver0_0_39;
